+++ front-end/src/dashboard/Dashboard.js
@@ -1,9 +1,9 @@
 import React, { useEffect, useState } from "react";
 import { useHistory } from "react-router-dom";
 import { listReservations } from "../utils/api";
-import Listing from "./Listing";
+import ReservationTable from "./reservationsCard/ReservationTable";
 import ErrorAlert from "../layout/ErrorAlert";
-import { next, previous } from "../utils/date-time";
+import CurrentDateToggle from "./reservationsCard/CurrentDateToggle";
 
 /**
  * Defines the dashboard page.
@@ -11,6 +11,15 @@ import { next, previous } from "../utils/date-time";
  *  the date for which the user wants to view reservations.
  * @returns {JSX.Element}
  */
+
+/* Edits to be made:
+      - Add Seat button to reservations linking to seat page
+      - Add Tables table with Free / Occupied tag
+        - Add load tables + availability function to loadDashboard
+      - CLEAN
+        - Create file for the card?
+ */
+
 function Dashboard({ date }) {
   const [reservations, setReservations] = useState([]);
   const [reservationsError, setReservationsError] = useState(null);
@@ -33,19 +42,7 @@ function Dashboard({ date }) {
       .then(setReservations)
       .catch(setReservationsError);
     return () => abortController.abort();
-  }
-
-  const goToPrev = () => {
-    setCurrentDate((current) => previous(current));
-  }
-
-  const goToToday = () => {
-    history.push("/");
-  }
-
-  const goToNext = () => {
-    setCurrentDate((current) => next(current));
-  }
+  }  
 
   return (
     <main>
@@ -57,26 +54,9 @@ function Dashboard({ date }) {
           </div>
           <div className="card-body">
             <ErrorAlert error={reservationsError} />
-            <table className="table table-striped">
-              <thead>
-                <tr>
-                  <th>Last Name</th>
-                  <th>First Name</th>
-                  <th>Mobile Number</th>
-                  <th>Time</th>
-                  <th>Party</th>
-                </tr>
-              </thead>
-              <tbody>
-                {reservations.map(reservation => <Listing key={`reservation-${reservation.reservation_id}`} reservation={reservation} />)}
-              </tbody>
-            </table>
+            <ReservationTable reservations={reservations} />
           </div>
-          <div className="card-footer">
-            <button type="button" className="btn btn-danger mr-3" onClick={goToPrev}>Prev</button>
-            <button type="button" className="btn btn-success mr-3" onClick={goToToday}>Today</button>
-            <button type="button" className="btn btn-primary" onClick={goToNext}>Next</button>
-          </div>
+          <CurrentDateToggle history={history} setCurrentDate={setCurrentDate} currentDate={currentDate}/>
         </div>
       </div>
